refactor(tui): move tool output history renderer
Extract preserved tool-output rendering, output row selection, and output wrapping helpers from history.rs into history/tool_output.rs while keeping OutputRow re-exported for the cache API.

Refs #3308.
