+++ crates/tui/src/tui/history.rs
@@ -23,6 +23,7 @@ mod constants;
 mod message;
 mod plan;
 mod thinking;
+mod tool_output;
 mod tool_run;
 
 use archived_context::{parse_archived_context, render_archived_context};
@@ -35,9 +36,8 @@ use checklist::{
 use checklist::{ChecklistChange, ChecklistItemSnapshot, ChecklistSnapshot};
 use constants::{
     ASSISTANT_GLYPH, TOOL_CARD_SUMMARY_LINES, TOOL_COMMAND_LINE_LIMIT, TOOL_DONE_SYMBOL,
-    TOOL_FAILED_SYMBOL, TOOL_HEADER_SUMMARY_LIMIT, TOOL_OUTPUT_HEAD_LINES, TOOL_OUTPUT_LINE_LIMIT,
-    TOOL_OUTPUT_TAIL_LINES, TOOL_RUNNING_SYMBOLS, TOOL_STATUS_SYMBOL_MS, TOOL_TEXT_LIMIT,
-    TRANSCRIPT_RAIL, USER_GLYPH,
+    TOOL_FAILED_SYMBOL, TOOL_HEADER_SUMMARY_LIMIT, TOOL_OUTPUT_LINE_LIMIT, TOOL_RUNNING_SYMBOLS,
+    TOOL_STATUS_SYMBOL_MS, TOOL_TEXT_LIMIT, TRANSCRIPT_RAIL, USER_GLYPH,
 };
 use message::{
     RenderedTranscriptLine, assistant_label_style_for, hard_break_copy_lines, message_body_style,
@@ -45,6 +45,7 @@ use message::{
     system_body_style, system_label_style, user_body_style, user_label_style,
 };
 use thinking::{render_hidden_thinking_activity, render_thinking};
+use tool_output::{render_exec_output_mode, render_tool_output_mode, wrap_plain_line, wrap_text};
 
 #[cfg(test)]
 use agent_activity::extract_agent_id;
@@ -57,6 +58,7 @@ pub use tool_run::{ToolRun, detect_tool_runs, detect_tool_runs_from_slices, tool
 
 #[cfg(test)]
 use thinking::{REASONING_CURSOR, REASONING_OPENER, REASONING_RAIL};
+pub use tool_output::OutputRow;
 
 use std::process::Command;
 
@@ -1818,15 +1820,6 @@ fn wrap_card_rail(mut lines: Vec<Line<'static>>) -> Vec<Line<'static>> {
     lines
 }
 
-fn render_tool_output_mode(
-    output: &str,
-    width: u16,
-    line_limit: usize,
-    mode: RenderMode,
-) -> Vec<Line<'static>> {
-    render_preserved_output_mode(output, width, line_limit, mode, "result")
-}
-
 fn review_severity_color(severity: &str) -> Color {
     match severity {
         "error" => palette::STATUS_ERROR,
@@ -1843,201 +1836,6 @@ fn format_review_location(path: Option<&String>, line: Option<u32>) -> String {
         (None, Some(line)) => format!("line {line}"),
         (None, None) => String::new(),
     }
-}
-
-fn render_exec_output_mode(
-    output: &str,
-    width: u16,
-    line_limit: usize,
-    mode: RenderMode,
-) -> Vec<Line<'static>> {
-    render_preserved_output_mode(output, width, line_limit, mode, "output")
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct OutputRow {
-    pub text: String,
-    pub intact: bool,
-}
-
-fn render_preserved_output_mode(
-    output: &str,
-    width: u16,
-    line_limit: usize,
-    mode: RenderMode,
-    first_label: &str,
-) -> Vec<Line<'static>> {
-    let mut lines = Vec::new();
-    if output.trim().is_empty() {
-        // #3031: In compact/Live mode, suppress "(no output)" — the tool
-        // header already carries the success/failure status.  Transcript
-        // mode still records it for exports/clipboard/pager.
-        if mode == RenderMode::Transcript {
-            lines.push(Line::from(Span::styled(
-                "  (no output)",
-                Style::default().fg(palette::TEXT_MUTED).italic(),
-            )));
-        }
-        return lines;
-    }
-
-    let content_hash = crate::tui::output_rows_cache::hash_str(output);
-    let all_lines = crate::tui::output_rows_cache::get_or_compute_rows(output, width, || {
-        output_rows(output, width)
-    });
-
-    if matches!(mode, RenderMode::Transcript) {
-        // Full-content path: emit every wrapped line with no head/tail split,
-        // no "+N more" affordance.
-        for (idx, row) in all_lines.iter().enumerate() {
-            render_output_row(
-                &mut lines,
-                if idx == 0 { Some(first_label) } else { None },
-                row,
-                width,
-            );
-        }
-        return lines;
-    }
-
-    let selected = crate::tui::output_rows_cache::get_or_compute_indices(
-        content_hash,
-        width,
-        line_limit,
-        || selected_output_indices(&all_lines, line_limit),
-    );
-    let mut previous: Option<usize> = None;
-    for (rendered_idx, idx) in selected.iter().copied().enumerate() {
-        if let Some(prev) = previous {
-            let omitted = idx.saturating_sub(prev + 1);
-            if omitted > 0 {
-                lines.push(details_affordance_line(
-                    &format!("{omitted} lines omitted"),
-                    Style::default().fg(palette::TEXT_MUTED),
-                ));
-            }
-        }
-
-        let row = &all_lines[idx];
-        render_output_row(
-            &mut lines,
-            if rendered_idx == 0 {
-                Some(first_label)
-            } else {
-                None
-            },
-            row,
-            width,
-        );
-        previous = Some(idx);
-    }
-
-    lines
-}
-
-fn output_rows(output: &str, width: u16) -> Vec<OutputRow> {
-    let wrap_width = width.saturating_sub(4).max(1) as usize;
-    let mut rows = Vec::new();
-    let mut sanitized = String::with_capacity(output.len());
-    for line in output.lines() {
-        sanitized.clear();
-        crate::tui::osc8::strip_ansi_into(line, &mut sanitized);
-        let intact = is_path_or_url_like(&sanitized);
-        if intact {
-            rows.push(OutputRow {
-                text: sanitized.clone(),
-                intact: true,
-            });
-        } else {
-            for wrapped in wrap_text(&sanitized, wrap_width) {
-                rows.push(OutputRow {
-                    text: wrapped,
-                    intact: false,
-                });
-            }
-        }
-    }
-    if rows.is_empty() {
-        rows.push(OutputRow {
-            text: String::new(),
-            intact: false,
-        });
-    }
-    rows
-}
-
-fn selected_output_indices(rows: &[OutputRow], line_limit: usize) -> Vec<usize> {
-    let total = rows.len();
-    if total <= line_limit || line_limit == 0 {
-        return (0..total).collect();
-    }
-
-    let head = TOOL_OUTPUT_HEAD_LINES.min(line_limit).min(total);
-    let tail = TOOL_OUTPUT_TAIL_LINES
-        .min(line_limit.saturating_sub(head))
-        .min(total.saturating_sub(head));
-    let mut selected = std::collections::BTreeSet::new();
-    selected.extend(0..head);
-    selected.extend(total.saturating_sub(tail)..total);
-
-    let budget = line_limit.saturating_sub(selected.len());
-    if budget > 0 {
-        let mut important: Vec<(usize, usize)> = rows
-            .iter()
-            .enumerate()
-            .skip(head)
-            .take(total.saturating_sub(head + tail))
-            .filter_map(|(idx, row)| output_importance_rank(&row.text).map(|rank| (idx, rank)))
-            .collect();
-        important.sort_by_key(|(idx, rank)| (*rank, *idx));
-        for (idx, _) in important.into_iter().take(budget) {
-            selected.insert(idx);
-        }
-    }
-
-    selected.into_iter().collect()
-}
-
-fn output_importance_rank(line: &str) -> Option<usize> {
-    let lower = line.to_ascii_lowercase();
-    if [
-        "error",
-        "failed",
-        "failure",
-        "fatal",
-        "panic",
-        "exception",
-        "traceback",
-        "denied",
-        "not found",
-        "no such file",
-        "cannot",
-        "can't",
-    ]
-    .iter()
-    .any(|needle| lower.contains(needle))
-    {
-        return Some(0);
-    }
-    if lower.contains("warning") || lower.contains("warn") {
-        return Some(1);
-    }
-    if is_path_or_url_like(line) {
-        return Some(2);
-    }
-    None
-}
-
-fn is_path_or_url_like(line: &str) -> bool {
-    let trimmed = line.trim();
-    if trimmed.contains("://") || trimmed.starts_with("file:") {
-        return true;
-    }
-    let has_separator = trimmed.contains('/') || trimmed.contains('\\');
-    let has_extension = trimmed
-        .split_whitespace()
-        .any(|part| part.rsplit_once('.').is_some_and(|(_, ext)| ext.len() <= 8));
-    has_separator && has_extension
 }
 
 /// Detect whether a system message is a cycle-boundary announcement
@@ -2074,121 +1872,6 @@ fn render_cycle_boundary(content: &str, width: u16) -> Vec<Line<'static>> {
         lines.push(Line::from(""));
     }
     lines
-}
-
-/// Detect whether a line contains a `path:line` pattern that could be
-/// opened by `try_open_file_at_line`. Returns a distinctive style
-/// (underline + blue) when the pattern matches, or `None` otherwise.
-/// The style is applied over the existing value style so the line
-/// remains readable.
-fn file_line_style(text: &str) -> Option<Style> {
-    let trimmed = text.trim();
-    if let Some((before, after)) = trimmed.rsplit_once(':')
-        && !before.is_empty()
-        && after.chars().all(|c| c.is_ascii_digit())
-        && looks_like_file_path(before)
-    {
-        Some(
-            Style::default()
-                .fg(palette::DEEPSEEK_SKY)
-                .add_modifier(Modifier::UNDERLINED),
-        )
-    } else {
-        None
-    }
-}
-
-/// Apply inline diff highlighting to a single text line.
-///
-/// Returns the appropriate style for the line based on its prefix:
-/// - Lines starting with `+` (after trimming) => `palette::DIFF_ADDED` (green)
-/// - Lines starting with `-` (after trimming) => `palette::STATUS_ERROR` (red)
-/// - Lines starting with `@@` => `palette::DEEPSEEK_SKY` (cyan/blue)
-/// - All other lines => None (use default style)
-fn diff_line_style(text: &str) -> Option<Style> {
-    let trimmed = text.trim_start();
-    if trimmed.starts_with("@@") {
-        Some(Style::default().fg(palette::WHALE_ACCENT_PRIMARY))
-    } else if trimmed.starts_with('+') && !trimmed.starts_with("+++") {
-        Some(Style::default().fg(palette::DIFF_ADDED))
-    } else if trimmed.starts_with('-') && !trimmed.starts_with("---") {
-        Some(Style::default().fg(palette::STATUS_ERROR))
-    } else {
-        None
-    }
-}
-
-fn render_output_row(
-    lines: &mut Vec<Line<'static>>,
-    label: Option<&str>,
-    row: &OutputRow,
-    width: u16,
-) {
-    // #374: apply file:line highlighting when the row text contains
-    // a `path:line` pattern. Diff style takes precedence (colored
-    // prefix lines should stay colored), but if no diff style matched,
-    // check for a file:line pattern and highlight it distinctively.
-    let diff_style = diff_line_style(&row.text);
-    let file_style = file_line_style(&row.text);
-    let value_style = diff_style.or(file_style).unwrap_or_else(tool_value_style);
-    if row.intact {
-        lines.push(render_card_detail_line_single(
-            label,
-            &row.text,
-            value_style,
-        ));
-    } else {
-        lines.extend(render_card_detail_line(
-            label,
-            &row.text,
-            value_style,
-            width,
-        ));
-    }
-}
-
-fn wrap_plain_line(line: &str, style: Style, width: u16) -> Vec<Line<'static>> {
-    let mut lines = Vec::new();
-    for part in wrap_text(line, width.max(1) as usize) {
-        lines.push(Line::from(Span::styled(part, style)));
-    }
-    lines
-}
-
-fn wrap_text(text: &str, width: usize) -> Vec<String> {
-    if width == 0 {
-        return vec![text.to_string()];
-    }
-    if text.is_empty() {
-        return vec![String::new()];
-    }
-
-    let mut lines = Vec::new();
-    let mut current = String::new();
-
-    for ch in text.chars() {
-        let tentative = if current.is_empty() {
-            ch.to_string()
-        } else {
-            let mut t = current.clone();
-            t.push(ch);
-            t
-        };
-
-        if UnicodeWidthStr::width(tentative.as_str()) > width && !current.is_empty() {
-            lines.push(std::mem::take(&mut current));
-        }
-
-        current.push(ch);
-    }
-
-    lines.push(current);
-
-    if lines.is_empty() {
-        vec![String::new()]
-    } else {
-        lines
-    }
 }
 
 fn status_symbol(started_at: Option<Instant>, status: ToolStatus, low_motion: bool) -> String {
